+++ frontend/src/components/creator-modal/creator-preview/creator-preview.scss
