+++ frontend/src/components/folder-modal/folder-modal.scss
